Fix 32-bit build: Add explicit casts and libatomic linkage
- Add static_cast for uint64_t to size_t conversions in timesamples-pprint.cc
- Add static_cast for atomic queue size calculations in task-queue.hh
- Link libatomic on 32-bit systems for 64-bit atomic operations

Fixes implicit conversion errors with -Wshorten-64-to-32 and undefined
reference to __atomic_load/__atomic_compare_exchange on i386.

+++ src/timesamples-pprint.cc
@@ -1815,7 +1815,7 @@ static void pprint_typed_array_timesamples(StreamWriter& writer, const PODTimeSa
             writer.write("None");
         } else {
             if (!samples._offsets.empty()) {
-                value_offset = samples._offsets[i];
+                value_offset = static_cast<size_t>(samples._offsets[i]);
             }
 
             // Get pointer to value data for this sample
@@ -1885,7 +1885,7 @@ void pprint_typed_array_timesamples<bool>(StreamWriter& writer, const PODTimeSam
             writer.write("None");
         } else {
             if (!samples._offsets.empty()) {
-                value_offset = samples._offsets[i];
+                value_offset = static_cast<size_t>(samples._offsets[i]);
             }
 
             // Get pointer to value data for this sample
